fix: use BIGINT for token count columns in PostgreSQL
Changed input_tokens, output_tokens, cache_read_tokens, and
cache_creation_tokens from INTEGER to BIGINT in usage_records and
audit_logs. This matches the Rust i64 types directly, removing the
need for i32→i64 conversions in From impls.

+++ crates/storage/migrations/postgres/20260415000000_initial.sql
@@ -71,8 +71,8 @@ CREATE TABLE IF NOT EXISTS usage_records (
     provider_id   TEXT NOT NULL,
     channel_id    TEXT,
     protocol      TEXT NOT NULL,
-    input_tokens  INTEGER,
-    output_tokens INTEGER,
+    input_tokens  BIGINT,
+    output_tokens BIGINT,
     cost          BIGINT NOT NULL,
     created_at    TIMESTAMP WITH TIME ZONE NOT NULL
 );
@@ -93,8 +93,8 @@ CREATE TABLE IF NOT EXISTS audit_logs (
     response_body TEXT NOT NULL,
     status_code   INTEGER NOT NULL,
     latency_ms    BIGINT NOT NULL,
-    input_tokens  INTEGER,
-    output_tokens INTEGER,
+    input_tokens  BIGINT,
+    output_tokens BIGINT,
     created_at    TIMESTAMP WITH TIME ZONE NOT NULL
 );
