+++ crates/storage/migrations/postgres/20260422000000_add_cache_read_prices_and_tokens.sql
@@ -1,3 +1,3 @@
 -- PostgreSQL migration: Add cache_read_price to channel_models and cache_read_tokens to usage_records
 ALTER TABLE channel_models ADD COLUMN IF NOT EXISTS cache_read_price REAL;
-ALTER TABLE usage_records ADD COLUMN IF NOT EXISTS cache_read_tokens INTEGER;
+ALTER TABLE usage_records ADD COLUMN IF NOT EXISTS cache_read_tokens BIGINT;
