+++ crates/storage/migrations/postgres/20260430000000_add_cache_creation_tokens.sql
@@ -1,2 +1,2 @@
 -- PostgreSQL migration: Add cache_creation_tokens to usage_records
-ALTER TABLE usage_records ADD COLUMN IF NOT EXISTS cache_creation_tokens INTEGER;
+ALTER TABLE usage_records ADD COLUMN IF NOT EXISTS cache_creation_tokens BIGINT;
